requirements.txt
streamlit
Pillow
boto3
pandas
langchain-aws==0.2.7
langchain_core==0.3.21